refactor(root): extract shared cli-args parser from its three script clones (#1238)
👤 For humans
The same tiny --flag parser lived in three harness scripts; it now lives
once in scripts/lib/cli-args.mjs, so a parsing fix lands everywhere.

🤖 For agents
- New scripts/lib/cli-args.mjs: parseArgs(argv, { boolean }) + isValue.
- eval-ledger/append.mjs: local parseArgs/rawFlags removed; --stdin/--check
  preserved via the boolean list; checkOnly now reads 'check' in args
  (same semantics as the old rawFlags probe).
- eval-ledger/from-pi-session.mjs + pi-box/apply-pi-config.mjs: local
  parseArgs/isValue removed, behavior identical.
- parseArgs carries a fallow-ignore-next-line complexity (CRAP inflated
  by 0 coverage; same precedent as scaffold-app.ts).
